PolicyManager/requirements.txt
cffi>=1.2.1
future>=0.15.2
gitdb>=0.6.4
GitPython>=1.0.1
lxml>=3.5.0
Cython>=0.23.4
# until fix for py3 installation has been commited in master branch, use benson-basis branch instead for CentOS
#pyjnius==1.4.0
#git clone https://github.com/benson-basis/pyjnius.git
# on OSX the master branch is good:
#git clone https://github.com/kivy/pyjnius.git
pyOpenSSL>=0.15.1
requests>=2.7.0
simplejson>=3.8.0
